requirements.txt
starlette==0.23.1
uvicorn==0.20.0
psycopg2-binary==2.9.5